+++ module/runtime/src/queue.ts
@@ -1,21 +1,19 @@
-import { Util } from '@travetto/runtime';
+import { Util } from './util';
 
 /**
- * WorkQueue, a manual async iterator.  Items are added manually, and consumed asynchronously
+ * An asynchronous queue
  */
-export class WorkQueue<X> implements AsyncIterator<X>, AsyncIterable<X> {
+export class AsyncQueue<X> implements AsyncIterator<X>, AsyncIterable<X> {
 
   #queue: X[] = [];
   #done = false;
   #ready = Util.resolvablePromise();
-  #size: number;
 
   /**
    * Initial set of items
    */
   constructor(initial: Iterable<X> = [], signal?: AbortSignal) {
     this.#queue.push(...initial);
-    this.#size = this.#queue.length;
     signal?.addEventListener('abort', () => this.close());
     if (signal?.aborted) {
       this.close();
@@ -44,18 +42,6 @@ export class WorkQueue<X> implements AsyncIterator<X>, AsyncIterable<X> {
    */
   add(item: X, immediate = false): void {
     this.#queue[immediate ? 'unshift' : 'push'](item);
-    this.#size += 1;
-    this.#ready.resolve();
-  }
-
-  /**
-   * Queue a list of data to stream
-   * @param {boolean} immediate Determines if item(s) should be append or prepended to the queue
-   */
-  addAll(items: Iterable<X>): void {
-    const copy = [...items];
-    this.#queue.push(...copy);
-    this.#size += copy.length;
     this.#ready.resolve();
   }
 
@@ -74,12 +60,5 @@ export class WorkQueue<X> implements AsyncIterator<X>, AsyncIterable<X> {
     this.#done = true;
     this.#ready.reject(e);
     return { value: undefined, done: this.#done };
-  }
-
-  /**
-   * Get size, will change as items are added
-   */
-  get size(): number {
-    return this.#size;
   }
 }
